Improve config files for validium
