--- infrastructure/zk/src/constants.ts
@@ -1,12 +1,0 @@
-export const VALIDIUM_COMPUTE_OVERHEAD_PART: number = 1.0;
-export const ROLLUP_COMPUTE_OVERHEAD_PART: number = 0.0;
-export const VALIDIUM_PUBDATA_OVERHEAD_PART: number = 0.0;
-export const ROLLUP_PUBDATA_OVERHEAD_PART: number = 1.0;
-export const VALIDIUM_BATCH_OVERHEAD_L1_GAS: number = 1000000;
-export const ROLLUP_BATCH_OVERHEAD_L1_GAS: number = 800000;
-export const VALIDIUM_MAX_PUBDATA_PER_BATCH: number = 1000000000000;
-export const ROLLUP_MAX_PUBDATA_PER_BATCH: number = 100000;
-export const VALIDIUM_L1_GAS_PER_PUBDATA_BYTE: number = 0;
-export const ROLLUP_L1_GAS_PER_PUBDATA_BYTE: number = 17;
-export const VALIDIUM_L1_BATCH_COMMIT_DATA_GENERATOR_MODE: string = '"Validium"';
-export const ROLLUP_L1_BATCH_COMMIT_DATA_GENERATOR_MODE: string = '"Rollup"';
