+++ infrastructure/zk/src/init.ts
@@ -14,14 +14,12 @@ import * as server from './server';
 import { up } from './up';
 
 import * as fs from 'fs';
-import * as constants from './constants';
 
 const entry = chalk.bold.yellow;
 const announce = chalk.yellow;
 const success = chalk.green;
 const timestamp = chalk.grey;
-const CHAIN_CONFIG_PATH = 'etc/env/base/chain.toml';
-const ETH_SENDER_PATH = 'etc/env/base/eth_sender.toml';
+
 
 export async function init(initArgs: InitArgs = DEFAULT_ARGS) {
     const {
@@ -35,7 +33,7 @@ export async function init(initArgs: InitArgs = DEFAULT_ARGS) {
 
     await announced(`Initializing in ${validiumMode ? 'Validium mode' : 'Roll-up mode'}`);
     process.env.VALIDIUM_MODE = validiumMode.toString();
-    await announced('Updating mode configuration', updateConfig(validiumMode));
+    await announced('Updating mode configuration', config.updateConfig(validiumMode));
     if (!process.env.CI && !skipEnvSetup) {
         await announced('Pulling images', docker.pull());
         await announced('Checking environment', checkEnv());
@@ -81,7 +79,7 @@ export async function init(initArgs: InitArgs = DEFAULT_ARGS) {
 export async function reinit(validiumMode: boolean) {
     process.env.VALIDIUM_MODE = validiumMode.toString();
     await announced(`Initializing in ${validiumMode ? 'Validium mode' : 'Roll-up mode'}`);
-    await announced('Updating mode configuration', updateConfig(validiumMode));
+    await announced('Updating mode configuration', config.updateConfig(validiumMode));
     await announced('Setting up containers', up());
     await announced('Compiling JS packages', run.yarn());
     await announced('Compile l2 contracts', compiler.compileAll());
@@ -104,7 +102,7 @@ export async function reinit(validiumMode: boolean) {
 export async function lightweightInit(validiumMode: boolean) {
     process.env.VALIDIUM_MODE = validiumMode.toString();
     await announced(`Initializing in ${validiumMode ? 'Validium mode' : 'Roll-up mode'}`);
-    await announced('Updating mode configuration', updateConfig(validiumMode));
+    await announced('Updating mode configuration', config.updateConfig(validiumMode));
     await announced(`Setting up containers`, up());
     await announced('Clean rocksdb', clean('db'));
     await announced('Clean backups', clean('backups'));
@@ -138,92 +136,6 @@ export async function announced(fn: string, promise: Promise<void> | void) {
 export async function submoduleUpdate() {
     await utils.exec('git submodule init');
     await utils.exec('git submodule update');
-}
-
-function updateConfigFile(path: string, modeConstantValues: Record<string, number | string | null>) {
-    let content = fs.readFileSync(path, 'utf-8');
-    let lines = content.split('\n');
-    let addedContent: string | undefined;
-    const lineIndices: Record<string, number> = {};
-
-    for (let i = 0; i < lines.length; i++) {
-        const line = lines[i];
-        if (!line.startsWith('#')) {
-            const match = line.match(/([^=]+)=(.*)/);
-            if (match) {
-                const key = match[1].trim();
-                lineIndices[key] = i;
-            }
-        }
-    }
-
-    for (const [key, value] of Object.entries(modeConstantValues)) {
-        const lineIndex = lineIndices[key];
-
-        if (lineIndex !== undefined) {
-            if (value !== null) {
-                lines.splice(lineIndex, 1, `${key}=${value}`);
-            } else {
-                lines.splice(lineIndex, 1);
-                for (const [k, index] of Object.entries(lineIndices)) {
-                    if (index > lineIndex) {
-                        lineIndices[k] = index - 1;
-                    }
-                }
-            }
-        } else {
-            if (value !== null) {
-                addedContent = `${key}=${value}\n`;
-            }
-        }
-    }
-
-    content = lines.join('\n');
-
-    if (addedContent) {
-        content += addedContent;
-    }
-
-    fs.writeFileSync(path, content);
-}
-function updateChainConfig(validiumMode: boolean) {
-    const modeConstantValues = {
-        compute_overhead_part: validiumMode
-            ? constants.VALIDIUM_COMPUTE_OVERHEAD_PART
-            : constants.ROLLUP_COMPUTE_OVERHEAD_PART,
-        pubdata_overhead_part: validiumMode
-            ? constants.VALIDIUM_PUBDATA_OVERHEAD_PART
-            : constants.ROLLUP_PUBDATA_OVERHEAD_PART,
-        batch_overhead_l1_gas: validiumMode
-            ? constants.VALIDIUM_BATCH_OVERHEAD_L1_GAS
-            : constants.ROLLUP_BATCH_OVERHEAD_L1_GAS,
-        max_pubdata_per_batch: validiumMode
-            ? constants.VALIDIUM_MAX_PUBDATA_PER_BATCH
-            : constants.ROLLUP_MAX_PUBDATA_PER_BATCH,
-        l1_batch_commit_data_generator_mode: validiumMode
-            ? constants.VALIDIUM_L1_BATCH_COMMIT_DATA_GENERATOR_MODE
-            : constants.ROLLUP_L1_BATCH_COMMIT_DATA_GENERATOR_MODE
-    };
-    updateConfigFile(CHAIN_CONFIG_PATH, modeConstantValues);
-}
-function updateEthSenderConfig(validiumMode: boolean) {
-    // This constant is used in validium mode and is deleted in rollup mode
-    // In order to pass the existing integration tests
-    const modeConstantValues = {
-        l1_gas_per_pubdata_byte: validiumMode
-            ? constants.VALIDIUM_L1_GAS_PER_PUBDATA_BYTE
-            : constants.ROLLUP_L1_GAS_PER_PUBDATA_BYTE
-    };
-    updateConfigFile(ETH_SENDER_PATH, modeConstantValues);
-}
-
-function updateConfig(validiumMode: boolean) {
-    updateChainConfig(validiumMode);
-    updateEthSenderConfig(validiumMode);
-    config.compileConfig();
-    let envFileContent = fs.readFileSync(process.env.ENV_FILE!).toString();
-    envFileContent += `VALIDIUM_MODE=${validiumMode}\n`;
-    fs.writeFileSync(process.env.ENV_FILE!, envFileContent);
 }
 
 async function checkEnv() {
